warning fixes

+++ tests/test_location_hash_query_bounding_box.cpp
@@ -6,7 +6,7 @@
 using namespace lochash;
 
 struct TestObject {
-	int         id;
+	size_t      id;
 	std::string name;
 };
 
@@ -103,7 +103,7 @@ TEST(LocationHelpersTest, QueryBoundingBoxComplexity)
 		locationHash.clear();
 		std::vector<TestObject> test_objects;
 		test_objects.reserve(count);
-		for (int i = 0; i < count; ++i) {
+		for (size_t i = 0; i < count; ++i) {
 			test_objects.push_back({i, "Object" + std::to_string(i)});
 			const float x = -1000.0f + static_cast<float>(rand()) / (static_cast<float>(RAND_MAX / 2000));
 			const float y = -1000.0f + static_cast<float>(rand()) / (static_cast<float>(RAND_MAX / 2000));
